refactor: 💡 major refactor and cleanup
