+++ config.yml
@@ -1,12 +1,12 @@
 game:
   players: 1 # valid values are 1, 2
   sound: true
-  music: true
+  music: false
   sound_vol: 1.0
   music_vol: 0.4
   soundtrack: "clouds.mp3"
-  show_grid: true
-  show_positions: true
+  show_grid: false
+  show_positions: false
   font_name: "clarity.ttf"
   grid_font_size: 10
   label_font_size: 20
