+++ knightfight/types.py
@@ -1,9 +1,8 @@
 from dataclasses import dataclass
 from enum import Enum
+import abc
 
 # enum for the piece types
-
-
 class PieceType(Enum):
     Bishop = "Bishop"
     King = "King"
@@ -37,3 +36,25 @@ class GridPosition:
         if isinstance(other, GridPosition):
             return self.row == other.row and self.col == other.col
         return False
+
+
+class State(abc.ABC):
+    """
+    Abstract class for the board state
+    """
+
+    @abc.abstractmethod
+    def get_cleared_state(self) -> dict:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def update_board_state(self) -> None:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def get_board_state(self) -> dict:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def get_starting_positions(self) -> dict:
+        raise NotImplementedError
